resume_doctor/requirements.txt
fastapi
uvicorn
google-generativeai
pdfminer.six
textstat
python-multipart
pydantic
python-dotenv
